added custom hook useSearchLineParams
