+++ hooks/use-search-params.ts
@@ -1,0 +1,31 @@
+import { usePathname, useRouter } from 'next/navigation';
+import qs from 'query-string';
+import { useEffect, useState } from 'react';
+
+export const useSearchLineParams = (searchParams: Record<string, string | number | null>) => {
+  const [isMounted, setIsMounted] = useState(false);
+
+  const pathname = usePathname();
+  const router = useRouter();
+
+  useEffect(() => {
+    setIsMounted(true);
+  }, []);
+
+  useEffect(() => {
+    const url = qs.stringifyUrl(
+      {
+        url: pathname,
+        query: searchParams,
+      },
+      { skipNull: true, skipEmptyString: true },
+    );
+
+    router.push(url);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [router, pathname, ...Object.values(searchParams)]);
+
+  if (!isMounted) {
+    return null;
+  }
+};
